@@ -11,12 +11,14 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.0.18",
+    "@radix-ui/react-icons": "^1.3.0",
     "@tailwindcss/forms": "^0.5.6",
     "lucide-react": "^0.284.0",
-    "next": "^13.5.5",
+    "next": "^14.0.1",
     "node-html-parser": "^6.1.10",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-iframe": "^1.8.5"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
