@@ -1,5 +1,5 @@
 import Article from "@/components/content/article";
-import settings from "@/fjord.json";
+import fjord from "@/fjord.config";
 import { notFound } from "next/navigation";
 
 type Post = {
@@ -35,7 +35,7 @@ type Post = {
 
 async function getPost(slug: string) {
   const res = await fetch(
-    `${settings.wp_url}/wp-json/wp/v2/posts?slug=${slug}&_embed`
+    `${fjord.wp_url}/wp-json/wp/v2/posts?slug=${slug}&_embed`
   );
 
   if (!res.ok) {
@@ -47,7 +47,7 @@ async function getPost(slug: string) {
 }
 
 export async function generateStaticParams() {
-  const res = await fetch(`${settings.wp_url}/wp-json/wp/v2/posts?_embed`);
+  const res = await fetch(`${fjord.wp_url}/wp-json/wp/v2/posts?_embed`);
 
   const data: Post[] = await res.json();
 
