@@ -1,4 +1,4 @@
-import settings from "@/fjord.json";
+import fjord from "@/fjord.config";
 import Link from "next/link";
 
 type Category = {
@@ -10,7 +10,7 @@ type Category = {
 export default async function CategoryPosts() {
   async function getCategoryPosts() {
     const res = await fetch(
-      `${settings.wp_url}/wp-json/wp/v2/categories?_embed&per_page=${settings.per_page}`
+      `${fjord.wp_url}/wp-json/wp/v2/categories?_embed&per_page=${fjord.per_page}`
     );
 
     if (!res.ok) {
