@@ -40,7 +40,7 @@ const Article = ({ post, date, author }: ArticleProps) => {
             dangerouslySetInnerHTML={{ __html: post.title.rendered }}
           ></h1>
           <div className="flex gap-2">
-            <p suppressHydrationWarning>{date.toDateString()}</p> |{author && <p>{author.name}</p>}
+            <p>{date.toDateString()}</p> |{author && <p>{author.name}</p>}
           </div>
           {post._embedded?.["wp:featuredmedia"] &&
             post._embedded["wp:featuredmedia"][0]?.media_details?.sizes?.full
