@@ -1,4 +1,4 @@
-import settings from "@/fjord.json";
+import fjord from "@/fjord.config";
 import Image from "next/image";
 import Link from "next/link";
 import SubscribeForm from "@/components/global/subscribe-form";
@@ -19,15 +19,29 @@ const Footer = () => {
   ];
 
   return (
-    <footer className="px-4 py-24 text-sm items-center m-auto bg-primary-300 grid gap-12"
-    >
+    <footer className="px-4 py-24 text-sm items-center m-auto bg-primary-300 dark:bg-primary-700 grid gap-12">
       <div className="px-6 max-w-7xl xl:px-0 grid md:grid-cols-3 gap-6 justify-between w-full m-auto">
         <div>
           <Link href="/">
-            <Image className="w-16 mb-4" width={200} height={200} src={settings.theme.logo} alt="logo"></Image>
+            <Image
+              priority
+              className="w-16 mb-4 dark:hidden"
+              width={200}
+              height={200}
+              src={fjord.theme.primary_logo}
+              alt="logo"
+            ></Image>
+            <Image
+              priority
+              className="w-16 mb-4 hidden dark:block"
+              width={200}
+              height={200}
+              src={fjord.theme.secondary_logo}
+              alt="logo"
+            ></Image>
           </Link>
-          <p className="text-sm text-secondary-500 md:w-3/4">
-            {settings.site_description}
+          <p className="text-sm text-secondary-400 md:w-3/4">
+            {fjord.site_description}
           </p>
         </div>
         <ul className="space-y-1">
@@ -46,7 +60,10 @@ const Footer = () => {
       </div>
       <div className="px-6 max-w-7xl xl:px-0 gap-6 flex-col-reverse md:flex-row flex justify-between w-full m-auto">
         <p suppressHydrationWarning>
-          © {new Date().getFullYear()} | {settings.site_name} | <a href="https://cameronyoungblood.com">Created</a> <a href="https://bridger.to">by</a> <a href="https://alpinecodex.com">Alpine Codex</a>
+          © {new Date().getFullYear()} | {fjord.site_name} |{" "}
+          <a href="https://cameronyoungblood.com">Created</a>{" "}
+          <a href="https://bridger.to">by</a>{" "}
+          <a href="https://alpinecodex.com">Alpine Codex</a>
         </p>
         <ul className="flex gap-4">
           {otherLinks.map((link, index) => (
